Menu and Settings page. And broke game into files
